fix(frontend): use drawers and detail tab nav on org settings
Replace centered modals with right-side drawers for API client creation,
secret reveal, and experiment/metric pickers. Align organization settings
tabs with the standard DetailTabNav detail-page layout.
